+++ tests/testthat.R
@@ -1,0 +1,4 @@
+library(testthat)
+library(BusinessReport)
+
+test_check("BusinessReport")